refactor: 프로젝트 디렉토리 수정
