+++ src/main/java/practice/Car.java
@@ -1,3 +1,5 @@
+package practice;
+
 public abstract class Car {
     /**
      * 리터당 이동 거리. 즉, 연비
