+++ src/main/java/practice/RentCompany.java
@@ -1,3 +1,5 @@
+package practice;
+
 import java.util.ArrayList;
 import java.util.List;
