+++ src/test/java/domain/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import domain.card.Card;
 import domain.card.CardNumber;
